fix: change stoi to stol for larger mc sample input

+++ src/main.cpp
@@ -19,7 +19,7 @@ int main(int argsc, const char* argsv[]) {
     }
     auto eu_options_list = load_options_from_csv(argsv[1]);
 
-    const uint32_t sample_count = argsc > 2 ? std::stoi(argsv[2]) : 1000000;
+    const uint32_t sample_count = argsc > 2 ? std::stol(argsv[2]) : 1000000;
     const monte_carlo_parameters params{sample_count};
 
     std::println("{} options // {} samples", eu_options_list.size(), params.sample_count);
